+++ tests/unit/plugins/modules/test_minio_policy.py
@@ -52,8 +52,8 @@ class TestMinioPolicy(ModuleTestCase):
                 {
                     "name": "testing",
                     "data": "{}",
-                    "access_key": "testing",
-                    "secret_key": "supersekret",
+                    "minio_access_key": "testing",
+                    "minio_secret_key": "supersekret",
                     "minio_url": "http://localhost:9000",
                 }
             )
@@ -72,8 +72,8 @@ class TestMinioPolicy(ModuleTestCase):
                 {
                     "name": "testing",
                     "data": "{}",
-                    "access_key": "testing",
-                    "secret_key": "supersekret",
+                    "minio_access_key": "testing",
+                    "minio_secret_key": "supersekret",
                     "minio_url": "http://localhost:9000",
                 }
             )
@@ -93,8 +93,8 @@ class TestMinioPolicy(ModuleTestCase):
                 {
                     "name": "testing",
                     "data": "{}",
-                    "access_key": "testing",
-                    "secret_key": "supersekret",
+                    "minio_access_key": "testing",
+                    "minio_secret_key": "supersekret",
                     "minio_url": "http://localhost:9000",
                 }
             )
@@ -119,8 +119,8 @@ class TestMinioPolicy(ModuleTestCase):
                 {
                     "name": "testing",
                     "data": "{}",
-                    "access_key": "testing",
-                    "secret_key": "supersekret",
+                    "minio_access_key": "testing",
+                    "minio_secret_key": "supersekret",
                     "minio_url": "http://localhost:9000",
                     "state": "absent",
                 }
@@ -146,8 +146,8 @@ class TestMinioPolicy(ModuleTestCase):
                 {
                     "name": "testing",
                     "data": "{}",
-                    "access_key": "testing",
-                    "secret_key": "supersekret",
+                    "minio_access_key": "testing",
+                    "minio_secret_key": "supersekret",
                     "minio_url": "http://localhost:9000",
                     "state": "absent",
                 }